+++ converters/html/src/html_visitor.rs
@@ -388,8 +388,7 @@ impl<W: Write> HtmlVisitor<W> {
         {
             writeln!(
                 self.writer,
-                r#"<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@fortawesome/fontawesome-free@7.2.0/css/all.min.css">
-<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@fortawesome/fontawesome-free@7.2.0/css/v4-shims.min.css">"#
+                r#"<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@fortawesome/fontawesome-free@7.2.0/css/all.min.css">"#
             )?;
         }